can zoom in and out
Tries to zoom in and out from mouse as center, but is really scaling
points from origin (which isn't bg_origin) and moving bg_origin (which
is more the position of the background) to keep the point where the
mouse is from moving relative to the screen. Should probably change it
to make sense.